% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spEigenCov.R
\name{spEigenCov}
\alias{spEigenCov}
\title{Sparse Spectral Decomposition of a Matrix}
\usage{
spEigenCov(S, q = 1, rho = 0.5, thres = 1e-09)
}
\arguments{
\item{S}{m-by-m covariance matrix.}

\item{q}{number of sparse eigenvectors.}

\item{rho}{sparsity weight factor. Any nonnegative number. Suggested range [0, 1].}

\item{thres}{threshold value. All the entries of the sparse eigenvectors less or equal to \code{thres} are set to 0. The default value is \code{1e-9}.}
}
\value{
A list with the following components:
\item{\code{vectors}  }{m-by-m matrix, columns corresponding to eigenvectors.}
\item{\code{values}  }{m-by-1 vector corresponding to eigenvalues.}
}
\description{
Computes sparse (orthogonal) eigenvectors of numeric covariance or data matrices.
}
\references{
K. Benidis, Y. Sun, P. Babu, D.P. Palomar "Orthogonal Sparse PCA and Covariance Estimation via Procrustes Reformulation,"
IEEE Transactions on Signal Processing, vol 64, no. 23, pp. 6211-6226, Dec. 2016.
}
\author{
Konstantinos Benidis and Daniel P. Palomar
}
