% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spEigen.R
\name{spEigen}
\alias{spEigen}
\title{Sparse Spectral Decomposition of a Matrix}
\usage{
spEigen(C, q, rho_nrm, d = NULL, V = NULL, thres = NULL)
}
\arguments{
\item{C}{n-by-m data matrix (n samples, m variables).}

\item{q}{number of eigenvectors to be estimated.}

\item{rho_nrm}{sparsity weight factor. Values from 0 to 1.}

\item{d}{1-by-q vector with weights.}

\item{V}{m-by-q initial point matrix. If not provided, the eigenvectors of the sample covariance matrix are used.}

\item{thres}{threshold value. All the entries of the sparse eigenvectors less or equal to \code{thres} are set to 0. The default value is \code{1e-9}.}
}
\value{
A list with the following components:
\item{\code{sp_vectors}  }{m-by-q matrix, columns corresponding to leading sparse eigenvectors.}
\item{\code{vectors}  }{m-by-q matrix, columns corresponding to leading eigenvectors.}
\item{\code{values}  }{q-by-1 vector corresponding to the leading eigenvalues.}
}
\description{
Computes sparse (orthogonal) eigenvectors of numeric data matrices.
}
\note{
If only the covariance matrix \code{S} is available and not the data matrix \code{C} then use Cholesky factorization and use the Cholesky factor instead of \code{C}.
}
\references{
K. Benidis, Y. Sun, P. Babu, D.P. Palomar "Orthogonal Sparse PCA and Covariance Estimation via Procrustes Reformulation,"
IEEE Transactions on Signal Processing, vol 64, no. 23, pp. 6211-6226, Dec. 2016.
}
\author{
Konstantinos Benidis, Daniel P. Palomar
}
